updated cft to make it work on startup

+++ install-factorio.sh
@@ -8,11 +8,17 @@ DIST_URL="https://www.factorio.com/get-download/latest/headless/linux64"
 TAR_FILE="$HOME/factorio-latest.tar.xz"
 INSTALL_DIR="/factorio"
 
+UPGRADE="$1"
+
 echo "Downloading archive from: $DIST_URL"
 wget "$DIST_URL" -O "$TAR_FILE"
 
-echo "Stopping factorio service..."
-systemctl stop factorio.service
+if [ UPGRADE == "upgrade"]
+then
+
+    echo "Stopping factorio service..."
+    systemctl stop factorio.service
+fi
 
 echo "Extracting archive..."
 tar -xf "$TAR_FILE" -C "$INSTALL_DIR" --strip-components=1
@@ -23,7 +29,10 @@ rm -f "$TAR_FILE"
 echo "Changing permissions..."
 chown -R factorio:factorio "$INSTALL_DIR"
 
-echo "Starting factorio service..."
-systemctl start factorio.service
+if [ UPGRADE == "upgrade"]
+then
+    echo "Starting factorio service..."
+    systemctl start factorio.service
+fi
 
 echo "Done."
